Finalize slides for ElixirDaze 2017
I actually removed all the code and talked about structure instead of
the code. I was told that the code was distracting so I will leave it
out. I still want to go back and make a solid example with tests to be
up with the slides.

Amos King @adkron <amos@binarynoggin.com>